Removes the function to return the InputStream / Outputstream. There is no reason to provide the function since the instances are being used immediately. This allows to have the for comprehension containing the instruction to create the input/output.
Removes the Open/Close state. The implementation requires specific methods to throw an exception. The new code relies on Option[ofstate]

Removes AuthResource.scala since it has 2 indirections to call new AwsS3ClientCreator.createS3Client. The trait is not used, and it creates an AuthResources to call aws.

Removes ThrowableEither.scala and moves the left side of either to throwable to avoid having the class.

Small improvements here and there avoiding redundant calls in for-comprehension